LayoutTests/imported/w3c/web-platform-tests/storage-access-api/requestStorageAccess-cross-site-sibling-iframes.sub.https.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unsupported permission name "storage-access".

Harness Error (TIMEOUT), message = null

TIMEOUT Grants have per-frame scope Test timed out
NOTRUN Cross-site sibling iframes should not be able to take advantage of the existing permission grant requested by others.
